# backend/app/api/contracts/step2-unified-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TMDB 海报墙",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/transfer/validate": {
      "post": {
        "tags": [
          "transfer"
        ],
        "summary": "验证分享链接",
        "description": "验证夸克分享链接是否有效\n\n返回链接中的文件列表",
        "operationId": "validate_link_api_transfer_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ValidateLinkData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transfer/exec": {
      "post": {
        "tags": [
          "transfer"
        ],
        "summary": "执行转存",
        "description": "执行转存操作\n\n- **collection_id**: 收藏 ID\n- **target_folder**: 目标目录 (可选，默认根据分类自动确定)\n- **auto_rename**: 是否自动重命名 (默认 False)",
        "operationId": "transfer_exec_api_transfer_exec_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TransferExecData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transfer/rename": {
      "post": {
        "tags": [
          "transfer"
        ],
        "summary": "独立重命名（SSE）",
        "operationId": "rename_collection_api_transfer_rename_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream. Each event payload is SseEnvelope JSON string prefixed with data: ",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_TransferExecData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/TransferExecData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[TransferExecData]"
      },
      "ApiResponse_ValidateLinkData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/ValidateLinkData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[ValidateLinkData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RenameRequest": {
        "properties": {
          "collection_id": {
            "type": "integer",
            "title": "Collection Id"
          }
        },
        "type": "object",
        "required": [
          "collection_id"
        ],
        "title": "RenameRequest",
        "description": "独立重命名请求"
      },
...
      "TransferExecData": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/TransferredFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "TransferExecData",
        "description": "执行转存数据"
      },
      "TransferExecRequest": {
        "properties": {
          "collection_id": {
            "type": "integer",
            "title": "Collection Id"
          },
          "target_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Folder"
          },
          "auto_rename": {
            "type": "boolean",
            "title": "Auto Rename",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "collection_id"
        ],
        "title": "TransferExecRequest",
        "description": "执行转存请求"
      },
      "TransferFile": {
        "properties": {
          "fid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 0
          },
          "is_dir": {
            "type": "boolean",
            "title": "Is Dir",
            "default": false
          }
        },
        "type": "object",
        "title": "TransferFile"
      },
...
      "TransferredFile": {
        "properties": {
          "fid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "type": "object",
        "title": "TransferredFile"
      },
      "ValidateLinkData": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/TransferFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "ValidateLinkData",
        "description": "验证链接数据"
      },
      "ValidateLinkRequest": {
        "properties": {
          "share_url": {
            "type": "string",
            "title": "Share Url"
          }
        },
        "type": "object",
        "required": [
          "share_url"
        ],
        "title": "ValidateLinkRequest",
        "description": "验证链接请求"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "SseEventData": {
        "type": "object",
        "required": [
          "type",
          "current",
          "total",
          "percentage",
          "message",
          "level"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "current": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "percentage": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "level": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "SseEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SseEventData"
          }
        }
      }
    }
  }
}